+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `--show-ip-addresses` flag to `server list` command to optionally include IP addresses in command output. 
 
 ## [1.4.0] - 2022-06-15
 ### Added